refactor: extract round-lifecycle / start_round cluster from state.py (#1271)
Next-increment extraction toward #1271 (state.py God-Object split),
stacked on the state-serialization cut (#1341).

Move the round-start / round-setup orchestration cluster out of
game/state.py into game/state_lifecycle.py as RoundLifecycleMixin:

  - start_game             (LOBBY -> PLAYING start gate)
  - start_round            (round-start orchestrator: song selection,
                            playback dispatch, metadata, round commit)
  - _ensure_media_player_service
  - _prepare_intro_round
  - _build_round_metadata
  - _initialize_round

Pure code move, API byte-for-byte unchanged: every caller hits these on
the GameState instance, which now inherits them via RoundLifecycleMixin.
GamePhase is lazy-imported inside the methods (# noqa: PLC0415) to keep
the import graph acyclic. Removed the now-unused get_song_uri /
ERR_GAME_* / MIN_PLAYERS imports from state.py.

Deliberately left on GameState (out of scope for this cut):
  - _score_all_players + _end_round_unlocked  (vote-window/scoring coupling)
  - _transition_to_reveal / _schedule_reveal_advance / _apply_reveal_lights
    (TTS / vote-window / party-lights coupling)
  - pause/resume, early-reveal, advance_to_end terminal helpers

state.py: 1818 -> 1586 lines (-232). Tests: 885 passed, 2 skipped.
ruff check + ruff format --check both clean.
